refactor: moved player health to PlayerEvents file
