+++ Assets/scripts/PlayerEvents.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32de0b1e590a0084bb76e9dc34a9743a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
